Move more things around, link everything up, update debug modules
