+++ packages/codec/lib/core/index.ts
@@ -5,12 +5,11 @@ export { readStack } from "../read/stack";
 export { slotAddress } from "../read/storage";
 export { StoragePointer } from "../types/pointer";
 export { ContractAllocationInfo, StorageAllocations, StorageMemberAllocation, AbiAllocations, CalldataAllocations, EventAllocations } from "../types/allocation";
-export { Slot, isWordsLength, equalSlots } from "../types/storage";
+export { Slot } from "../types/storage";
 export { DecoderRequest } from "../types/request";
 export { EvmInfo, AllocationInfo } from "../types/evm";
 export { CalldataDecoding, LogDecoding } from "../types/decoding";
 
 export { decodeVariable, decodeEvent, decodeCalldata } from "./decoding";
-export { abifyCalldataDecoding, abifyLogDecoding } from "./abify";
 
 export { encodeAbi, encodeTupleAbi } from "../encode/abi"; //we have no need to export these at present, but someone might want them?
