+++ packages/codec/lib/decode/event.ts
@@ -3,7 +3,8 @@ const debug = debugModule("codec:decode:event");
 
 import decodeValue from "./value";
 import read from "../read";
-import { Types, Values, Errors, Conversion as ConversionUtils } from "truffle-codec-utils";
+import { Types, Values, Errors } from "../format";
+import { Conversion as ConversionUtils, TypeUtils } from "../utils";
 import { EventTopicPointer } from "../types/pointer";
 import { EvmInfo } from "../types/evm";
 import { DecoderOptions } from "../types/options";
@@ -11,7 +12,7 @@ import { DecoderRequest } from "../types/request";
 import { StopDecodingError } from "../types/errors";
 
 export default function* decodeTopic(dataType: Types.Type, pointer: EventTopicPointer, info: EvmInfo, options: DecoderOptions = {}): Generator<DecoderRequest, Values.Result, Uint8Array> {
-  if(Types.isReferenceType(dataType)) {
+  if(TypeUtils.isReferenceType(dataType)) {
     //we cannot decode reference types "stored" in topics; we have to just return an error
     let bytes: Uint8Array = yield* read(pointer, info.state);
     let raw: string = ConversionUtils.toHexString(bytes);
