+++ packages/codec/lib/format/types.ts
@@ -1,5 +1,5 @@
 import debugModule from "debug";
-const debug = debugModule("codec-utils:types:types");
+const debug = debugModule("codec:format:types");
 
 //type objects for Solidity types
 //these will just be defined as interfaces; there's not any particular need for
@@ -15,10 +15,7 @@ const debug = debugModule("codec-utils:types:types");
 //that might change in the future.
 
 import BN from "bn.js";
-import * as Ast from "../ast";
-import { Definition as DefinitionUtils } from "../definition";
-import { CompilerVersion, solidityFamily } from "../compiler";
-import { AbiUtils } from "../abi";
+import * as Ast from "../types/ast";
 
 export namespace Types {
 
@@ -256,622 +253,5 @@ export namespace Types {
   export interface TypesById {
     [id: string]: UserDefinedType;
   };
-
-  //NOTE: the following function will *not* work for arbitrary nodes! It will,
-  //however, work for the ones we need (i.e., variable definitions, and arbitrary
-  //things of elementary type)
-  //NOTE: set forceLocation to *null* to force no location. leave it undefined
-  //to not force a location.
-  export function definitionToType(definition: Ast.AstDefinition, compiler: CompilerVersion, forceLocation?: Ast.Location | null): Type {
-    debug("definition %O", definition);
-    let typeClass = DefinitionUtils.typeClass(definition);
-    let typeHint = DefinitionUtils.typeStringWithoutLocation(definition);
-    switch(typeClass) {
-      case "bool":
-        return {
-          typeClass,
-          typeHint
-        };
-      case "address": {
-        switch(solidityFamily(compiler)) {
-          case "pre-0.5.0":
-            return {
-              typeClass,
-              kind: "general",
-              typeHint
-            };
-          case "0.5.x":
-            return {
-              typeClass,
-              kind: "specific",
-              payable: DefinitionUtils.typeIdentifier(definition) === "t_address_payable"
-            };
-        }
-        break; //to satisfy typescript
-      }
-      case "uint": {
-        let bytes = DefinitionUtils.specifiedSize(definition);
-        return {
-          typeClass,
-          bits: bytes * 8,
-          typeHint
-        };
-      }
-      case "int": { //typeScript won't let me group these for some reason
-        let bytes = DefinitionUtils.specifiedSize(definition);
-        return {
-          typeClass,
-          bits: bytes * 8,
-          typeHint
-        };
-      }
-      case "fixed": { //typeScript won't let me group these for some reason
-        let bytes = DefinitionUtils.specifiedSize(definition);
-        let places = DefinitionUtils.decimalPlaces(definition);
-        return {
-          typeClass,
-          bits: bytes * 8,
-          places,
-          typeHint
-        };
-      }
-      case "ufixed": {
-        let bytes = DefinitionUtils.specifiedSize(definition);
-        let places = DefinitionUtils.decimalPlaces(definition);
-        return {
-          typeClass,
-          bits: bytes * 8,
-          places,
-          typeHint
-        };
-      }
-      case "string": {
-        if(forceLocation === null) {
-          return {
-            typeClass,
-            typeHint
-          };
-        }
-        let location = forceLocation || DefinitionUtils.referenceType(definition);
-        return {
-          typeClass,
-          location,
-          typeHint
-        };
-      }
-      case "bytes": {
-        let length = DefinitionUtils.specifiedSize(definition);
-        if(length !== null) {
-          return {
-            typeClass,
-            kind: "static",
-            length,
-            typeHint
-          }
-        } else {
-          if(forceLocation === null) {
-            return {
-              typeClass,
-              kind: "dynamic",
-              typeHint
-            };
-          }
-          let location = forceLocation || DefinitionUtils.referenceType(definition);
-          return {
-            typeClass,
-            kind: "dynamic",
-            location,
-            typeHint
-          }
-        }
-      }
-      case "array": {
-        let baseDefinition = DefinitionUtils.baseDefinition(definition);
-        let baseType = definitionToType(baseDefinition, compiler, forceLocation);
-        let location = forceLocation || DefinitionUtils.referenceType(definition);
-        if(DefinitionUtils.isDynamicArray(definition)) {
-          if(forceLocation !== null) {
-            return {
-              typeClass,
-              baseType,
-              kind: "dynamic",
-              location,
-              typeHint
-            }
-          }
-          else {
-            return {
-              typeClass,
-              baseType,
-              kind: "dynamic",
-              typeHint
-            }
-          }
-        } else {
-          let length = new BN(DefinitionUtils.staticLengthAsString(definition));
-          if(forceLocation !== null) {
-            return {
-              typeClass,
-              baseType,
-              kind: "static",
-              length,
-              location,
-              typeHint
-            }
-          }
-          else {
-            return {
-              typeClass,
-              baseType,
-              kind: "static",
-              length,
-              typeHint
-            }
-          }
-        }
-      }
-      case "mapping": {
-        let keyDefinition = DefinitionUtils.keyDefinition(definition);
-        //note that we can skip the scopes argument here! that's only needed when
-        //a general node, rather than a declaration, is being passed in
-        let keyType = <ElementaryType>definitionToType(keyDefinition, compiler, null);
-        //suppress the location on the key type (it'll be given as memory but
-        //this is meaningless)
-        //also, we have to tell TypeScript ourselves that this will be an elementary
-        //type; it has no way of knowing that
-        let valueDefinition = definition.valueType || definition.typeName.valueType;
-        let valueType = definitionToType(valueDefinition, compiler, forceLocation);
-        if(forceLocation === null) {
-          return {
-            typeClass,
-            keyType,
-            valueType
-          };
-        }
-        return {
-          typeClass,
-          keyType,
-          valueType,
-          location: "storage"
-        };
-      }
-      case "function": {
-        let visibility = DefinitionUtils.visibility(definition);
-        let mutability = DefinitionUtils.mutability(definition);
-        let [inputParameters, outputParameters] = DefinitionUtils.parameters(definition);
-        //note: don't force a location on these! use the listed location!
-        let inputParameterTypes = inputParameters.map(parameter => definitionToType(parameter, compiler));
-        let outputParameterTypes = outputParameters.map(parameter => definitionToType(parameter, compiler));
-        switch(visibility) {
-          case "internal":
-            return {
-              typeClass,
-              visibility,
-              mutability,
-              inputParameterTypes,
-              outputParameterTypes
-            };
-          case "external":
-            return {
-              typeClass,
-              visibility,
-              kind: "specific",
-              mutability,
-              inputParameterTypes,
-              outputParameterTypes
-            };
-        }
-        break; //to satisfy typescript
-      }
-      case "struct": {
-        let id = DefinitionUtils.typeId(definition).toString();
-        let qualifiedName = DefinitionUtils.typeStringWithoutLocation(definition).match(/struct (.*)/)[1];
-        let [definingContractName, typeName] = qualifiedName.split(".");
-        if(forceLocation === null) {
-          return {
-            typeClass,
-            kind: "local",
-            id,
-            typeName,
-            definingContractName
-          };
-        }
-        let location = forceLocation || DefinitionUtils.referenceType(definition);
-        return {
-          typeClass,
-          kind: "local",
-          id,
-          typeName,
-          definingContractName,
-          location
-        };
-      }
-      case "enum": {
-        let id = DefinitionUtils.typeId(definition).toString();
-        let qualifiedName = DefinitionUtils.typeStringWithoutLocation(definition).match(/enum (.*)/)[1];
-        let [definingContractName, typeName] = qualifiedName.split(".");
-        return {
-          typeClass,
-          kind: "local",
-          id,
-          typeName,
-          definingContractName
-        };
-      }
-      case "contract": {
-        let id = DefinitionUtils.typeId(definition).toString();
-        let typeName = definition.typeName
-          ? definition.typeName.name
-          : definition.name;
-        let contractKind = DefinitionUtils.contractKind(definition);
-        return {
-          typeClass,
-          kind: "native",
-          id,
-          typeName,
-          contractKind
-        }
-      }
-      case "magic": {
-        let typeIdentifier = DefinitionUtils.typeIdentifier(definition);
-        let variable = <MagicVariableName> typeIdentifier.match(/^t_magic_(.*)$/)[1];
-        return {
-          typeClass,
-          variable
-        }
-      }
-    }
-  }
-
-  //whereas the above takes variable definitions, this takes the actual type
-  //definition
-  export function definitionToStoredType(definition: Ast.AstDefinition, compiler: CompilerVersion, referenceDeclarations?: Ast.AstReferences): UserDefinedType {
-    switch(definition.nodeType) {
-      case "StructDefinition": {
-        let id = definition.id.toString();
-        let [definingContractName, typeName] = definition.canonicalName.split(".");
-        let memberTypes: {name: string, type: Type}[] = definition.members.map(
-          member => ({name: member.name, type: definitionToType(member, compiler, null)})
-        );
-        let definingContract;
-        if(referenceDeclarations) {
-          let contractDefinition = Object.values(referenceDeclarations).find(
-            node => node.nodeType === "ContractDefinition" &&
-            node.nodes.some(
-              (subNode: Ast.AstDefinition) => subNode.id.toString() === id
-            )
-          );
-          definingContract = <ContractTypeNative> definitionToStoredType(contractDefinition, compiler); //can skip reference declarations
-        }
-        return {
-          typeClass: "struct",
-          kind: "local",
-          id,
-          typeName,
-          definingContractName,
-          definingContract,
-          memberTypes
-        };
-      }
-      case "EnumDefinition": {
-        let id = definition.id.toString();
-        let [definingContractName, typeName] = definition.canonicalName.split(".");
-        let options = definition.members.map(member => member.name);
-        let definingContract;
-        if(referenceDeclarations) {
-          let contractDefinition = Object.values(referenceDeclarations).find(
-            node => node.nodeType === "ContractDefinition" &&
-            node.nodes.some(
-              (subNode: Ast.AstDefinition) => subNode.id.toString() === id
-            )
-          );
-          definingContract = <ContractTypeNative> definitionToStoredType(contractDefinition, compiler); //can skip reference declarations
-        }
-        return {
-          typeClass: "enum",
-          kind: "local",
-          id,
-          typeName,
-          definingContractName,
-          definingContract,
-          options
-        };
-      }
-      case "ContractDefinition": {
-        let id = definition.id.toString();
-        let typeName = definition.name;
-        let contractKind = definition.contractKind;
-        let payable = DefinitionUtils.isContractPayable(definition);
-        return {
-          typeClass: "contract",
-          kind: "native",
-          id,
-          typeName,
-          contractKind,
-          payable
-        };
-      }
-    }
-  }
-
-  //one could define a counterpart function that stripped all unnecessary information
-  //from the type object, but at the moment I see no need for that
-  export function fullType(basicType: Type, userDefinedTypes: TypesById): Type {
-    if(!isUserDefinedType(basicType)) {
-      return basicType;
-    }
-    let id = basicType.id;
-    let storedType = userDefinedTypes[id];
-    if(!storedType) {
-      return basicType;
-    }
-    let returnType: Type = { ...basicType, ...storedType };
-    if(isReferenceType(basicType) && basicType.location !== undefined) {
-      returnType = specifyLocation(returnType, basicType.location);
-    }
-    return returnType;
-  }
-
-  export function isContractDefinedType(anyType: Type): anyType is ContractDefinedType {
-    const contractDefinedTypes = ["enum", "struct"];
-    return contractDefinedTypes.includes(anyType.typeClass);
-  }
-
-  export function isUserDefinedType(anyType: Type): anyType is UserDefinedType {
-    const userDefinedTypes = ["contract", "enum", "struct"];
-    return userDefinedTypes.includes(anyType.typeClass);
-  }
-
-  export function isReferenceType(anyType: Type): anyType is ReferenceType {
-    const alwaysReferenceTypes = ["array", "mapping", "struct", "string"];
-    if(alwaysReferenceTypes.includes(anyType.typeClass)) {
-      return true;
-    }
-    else if(anyType.typeClass === "bytes") {
-      return anyType.kind === "dynamic";
-    }
-    else {
-      return false;
-    }
-  }
-
-  //the location argument here always forces, so passing undefined *will* force undefined
-  export function specifyLocation(dataType: Type, location: Ast.Location | undefined): Type {
-    if(isReferenceType(dataType)) {
-      switch(dataType.typeClass) {
-        case "string":
-        case "bytes":
-          return { ...dataType, location };
-        case "array":
-          return {
-            ...dataType,
-            location,
-            baseType: specifyLocation(dataType.baseType, location)
-          };
-        case "mapping":
-          let newLocation = location === "storage" ? "storage" as "storage" : undefined;
-          return {
-            ...dataType,
-            location: newLocation,
-            valueType: specifyLocation(dataType.valueType, newLocation)
-          };
-        case "struct":
-          let returnType = { ...dataType, location };
-          if(returnType.memberTypes) {
-            returnType.memberTypes = returnType.memberTypes.map(
-              ({name: memberName, type: memberType}) => ({name: memberName, type: specifyLocation(memberType, location)})
-            );
-          }
-          return returnType;
-      }
-    }
-    else {
-      return dataType;
-    }
-  }
-
-  export function abiParameterToType(abi: AbiUtils.AbiParameter): Type {
-    let typeName = abi.type;
-    let typeHint = abi.internalType;
-    //first: is it an array?
-    let arrayMatch = typeName.match(/(.*)\[(\d*)\]$/);
-    if(arrayMatch) {
-      let baseTypeName = arrayMatch[1];
-      let lengthAsString = arrayMatch[2]; //may be empty!
-      let baseAbi = {...abi, type: baseTypeName};
-      let baseType = abiParameterToType(baseAbi);
-      if(lengthAsString === "") {
-        return {
-          typeClass: "array",
-          kind: "dynamic",
-          baseType,
-          typeHint
-        }
-      }
-      else {
-        let length = new BN(lengthAsString);
-        return {
-          typeClass: "array",
-          kind: "static",
-          length,
-          baseType,
-          typeHint
-        }
-      }
-    }
-    //otherwise, here are the simple cases
-    let typeClass = typeName.match(/^([^0-9]+)/)[1];
-    switch(typeClass) {
-      case "uint":
-      case "int": {
-          let bits = typeName.match(/^u?int([0-9]+)/)[1];
-          return {
-            typeClass,
-            bits: parseInt(bits),
-            typeHint
-          };
-      }
-      case "bytes":
-        let length = typeName.match(/^bytes([0-9]*)/)[1];
-        if(length === "") {
-          return {
-            typeClass,
-            kind: "dynamic",
-            typeHint
-          };
-        }
-        else {
-          return {
-            typeClass,
-            kind: "static",
-            length: parseInt(length),
-            typeHint
-          };
-        }
-      case "address":
-        return {
-          typeClass,
-          kind: "general",
-          typeHint
-        };
-      case "string":
-      case "bool":
-        return {
-          typeClass,
-          typeHint
-        };
-      case "fixed":
-      case "ufixed": {
-          let [_, bits, places] = typeName.match(/^u?fixed([0-9]+)x([0-9]+)/);
-          return {
-            typeClass,
-            bits: parseInt(bits),
-            places: parseInt(places),
-            typeHint
-          };
-      }
-      case "function":
-        return {
-          typeClass,
-          visibility: "external",
-          kind: "general",
-          typeHint
-        }
-      case "tuple":
-        let memberTypes = abi.components.map(
-          component => ({
-            name: component.name || undefined, //leave undefined if component.name is empty string
-            type: abiParameterToType(component)
-          })
-        );
-        return {
-          typeClass,
-          memberTypes,
-          typeHint
-        };
-    }
-  }
-
-  //NOTE: the following two functions might not be exactly right for weird internal stuff,
-  //or for ABI-only stuff.  (E.g. for internal stuff sometimes it records whether things
-  //are pointers or not??  we don't track that so we can't recreate that)
-  //But what can you do.
-
-  export function typeString(dataType: Type): string {
-    let baseString = typeStringWithoutLocation(dataType);
-    if(isReferenceType(dataType) && dataType.location) {
-      return baseString + " " + dataType.location;
-    }
-    else {
-      return baseString;
-    }
-  }
-
-  export function typeStringWithoutLocation(dataType: Type): string {
-    switch(dataType.typeClass) {
-      case "uint":
-        return dataType.typeHint || `uint${dataType.bits}`;
-      case "int":
-        return dataType.typeHint || `int${dataType.bits}`;
-      case "bool":
-        return dataType.typeHint || "bool";
-      case "bytes":
-        if(dataType.typeHint) {
-          return dataType.typeHint;
-        }
-        switch(dataType.kind) {
-          case "dynamic":
-            return "bytes";
-          case "static":
-            return `bytes${dataType.length}`;
-          }
-      case "address":
-        switch(dataType.kind) {
-          case "general":
-            return dataType.typeHint || "address"; //I guess?
-          case "specific":
-            return dataType.payable ? "address payable" : "address";
-        }
-      case "string":
-        return dataType.typeHint || "string";
-      case "fixed":
-        return dataType.typeHint || `fixed${dataType.bits}x${dataType.places}`;
-      case "ufixed":
-        return dataType.typeHint || `ufixed${dataType.bits}x${dataType.places}`;
-      case "array":
-        if(dataType.typeHint) {
-          return dataType.typeHint;
-        }
-        switch(dataType.kind) {
-          case "dynamic":
-            return `${typeStringWithoutLocation(dataType.baseType)}[]`;
-          case "static":
-            return `${typeStringWithoutLocation(dataType.baseType)}[${dataType.length}]`;
-        }
-      case "mapping":
-        return `mapping(${typeStringWithoutLocation(dataType.keyType)} => ${typeStringWithoutLocation(dataType.valueType)})`;
-      case "struct":
-      case "enum":
-        //combining these cases for simplicity
-        switch(dataType.kind) {
-          case "local":
-            return `${dataType.typeClass} ${dataType.definingContractName}.${dataType.typeName}`;
-          case "global": //WARNING, SPECULATIVE
-            return `${dataType.typeClass} ${dataType.typeName}`;
-        }
-      case "tuple":
-        return dataType.typeHint || "tuple(" + dataType.memberTypes.map(memberType => typeString(memberType.type)).join(",") + ")"; //note that we do include location and do not put spaces
-      case "contract":
-        return dataType.contractKind + " " + dataType.typeName;
-      case "magic":
-        //no, this is not transposed!
-        const variableNames = {message: "msg", transaction: "tx", block: "block"};
-        return variableNames[dataType.variable];
-      case "function":
-        let visibilityString: string;
-        switch(dataType.visibility) {
-          case "external":
-            if(dataType.kind === "general") {
-              if(dataType.typeHint) {
-                return dataType.typeHint;
-              }
-              else {
-                return "function external"; //I guess???
-              }
-            }
-            visibilityString = " external"; //note the deliberate space!
-            break;
-          case "internal":
-            visibilityString = "";
-            break;
-        }
-        let mutabilityString = dataType.mutability === "nonpayable" ? " " + dataType.mutability : ""; //again, note the deliberate space
-        let inputList = dataType.inputParameterTypes.map(typeString).join(","); //note that we do include location, and do not put spaces
-        let outputList = dataType.outputParameterTypes.map(typeString).join(",");
-        let inputString = `function(${inputList})`;
-        let outputString = outputList === "" ? "" : ` returns (${outputList})`; //again, note the deliberate space
-        return inputString + mutabilityString + visibilityString + outputString;
-    }
-  }
 
 }
