+++ packages/codec/lib/read/stack.ts
@@ -1,11 +1,12 @@
 import debugModule from "debug";
 const debug = debugModule("codec:read:stack");
 
-import * as CodecUtils from "truffle-codec-utils";
+import * as CodecUtils from "../utils";
+import { Errors } from "../format/errors";
 
 export function readStack(stack: Uint8Array[], from: number, to: number): Uint8Array {
   if(from < 0 || to >= stack.length) {
-    throw new CodecUtils.Errors.DecodingError(
+    throw new Errors.DecodingError(
       {
         kind: "ReadErrorStack",
         from,
