+++ packages/codec/lib/utils/abi.ts
@@ -1,10 +1,11 @@
 import debugModule from "debug";
-const debug = debugModule("codec-utils:abi");
+const debug = debugModule("codec:utils:abi");
 
 import { Abi as SchemaAbi } from "truffle-contract-schema/spec";
 import { EVM as EVMUtils } from "./evm";
-import { AstDefinition, AstReferences, Mutability } from "./ast";
+import { AstDefinition, AstReferences, Mutability } from "../types/ast";
 import { definitionToAbi } from "./definition2abi";
+import * as AbiTypes from "../types/abi";
 import Web3 from "web3";
 
 //NOTE: SchemaAbi is kind of loose and a pain to use.
@@ -14,106 +15,60 @@ import Web3 from "web3";
 
 export namespace AbiUtils {
 
-  export type Abi = AbiEntry[];
-
-  export type AbiEntry = FunctionAbiEntry | ConstructorAbiEntry | FallbackAbiEntry | EventAbiEntry;
-
-  export interface FunctionAbiEntry {
-    type: "function";
-    name: string;
-    inputs: AbiParameter[];
-    outputs: AbiParameter[];
-    stateMutability?: Mutability; //only in newer ones
-    constant?: boolean; //only in older ones
-    payable?: boolean; //only in older ones
-  }
-
-  export interface ConstructorAbiEntry {
-    type: "constructor";
-    inputs: AbiParameter[];
-    stateMutability?: "payable" | "nonpayable"; //only in newer ones
-    payable?: boolean; //only in older ones
-  }
-
-  export interface FallbackAbiEntry {
-    type: "fallback";
-    stateMutability?: "payable" | "nonpayable"; //only in newer ones
-    payable?: boolean; //only in older ones
-  }
-
-  export interface EventAbiEntry {
-    type: "event";
-    name: string;
-    inputs: AbiParameter[];
-    anonymous: boolean;
-  }
-
-  export interface AbiParameter {
-    name: string;
-    type: string;
-    indexed?: boolean; //only present for inputs
-    components?: AbiParameter[]; //only preset for tuples (structs)
-    internalType?: string;
-  }
-
-  export interface FunctionAbiBySelectors {
-    [selector: string]: FunctionAbiEntry
-  }
-
-  export const DEFAULT_CONSTRUCTOR_ABI: ConstructorAbiEntry = {
+  export const DEFAULT_CONSTRUCTOR_ABI: AbiTypes.ConstructorAbiEntry = {
     type: "constructor",
     inputs: [],
     stateMutability: "nonpayable",
     payable: false
   };
 
-  export const DEFAULT_FALLBACK_ABI: FallbackAbiEntry = {
+  export const DEFAULT_FALLBACK_ABI: AbiTypes.FallbackAbiEntry = {
     type: "fallback",
     stateMutability: "nonpayable",
     payable: false
   };
 
-  export function schemaAbiToAbi(abiLoose: SchemaAbi): Abi {
+  export function schemaAbiToAbi(abiLoose: SchemaAbi): AbiTypes.Abi {
     return abiLoose.map(
       entry => entry.type
-        ? <AbiEntry> entry
-        : <AbiEntry> {type: "function", ...entry}
+        ? <AbiTypes.AbiEntry> entry
+        : <AbiTypes.AbiEntry> {type: "function", ...entry}
     );
   }
 
   //note the return value only includes functions!
-  export function computeSelectors(abi: Abi | undefined): FunctionAbiBySelectors | undefined {
+  export function computeSelectors(abi: AbiTypes.Abi | undefined): AbiTypes.FunctionAbiBySelectors | undefined {
     if(abi === undefined) {
       return undefined;
     }
     return Object.assign({},
       ...abi.filter(
-        (abiEntry: AbiEntry) => abiEntry.type === "function"
+        (abiEntry: AbiTypes.AbiEntry) => abiEntry.type === "function"
       ).map(
-        (abiEntry: FunctionAbiEntry) => ({ [abiSelector(abiEntry)]: abiEntry })
+        (abiEntry: AbiTypes.FunctionAbiEntry) => ({ [abiSelector(abiEntry)]: abiEntry })
       )
     )
   }
 
   //does this ABI have a payable fallback function?
-  export function abiHasPayableFallback(abi: Abi | undefined): boolean | undefined {
+  export function abiHasPayableFallback(abi: AbiTypes.Abi | undefined): boolean | undefined {
     if(abi === undefined) {
       return undefined;
     }
     return abiMutability(getFallbackEntry(abi)) === "payable";
   }
 
-  export function abiHasFallback(abi: Abi) {
-    return abi.some((abiEntry: AbiEntry) => abiEntry.type === "fallback");
+  export function abiHasFallback(abi: AbiTypes.Abi) {
+    return abi.some((abiEntry: AbiTypes.AbiEntry) => abiEntry.type === "fallback");
   }
 
   //gets the fallback entry; if there isn't one, returns a default one
-  export function getFallbackEntry(abi: Abi): FallbackAbiEntry {
+  export function getFallbackEntry(abi: AbiTypes.Abi): AbiTypes.FallbackAbiEntry {
     //no idea why TS's type inference is failing on this one...
-    return <FallbackAbiEntry> abi.find(abiEntry => abiEntry.type === "fallback") || DEFAULT_FALLBACK_ABI;
+    return <AbiTypes.FallbackAbiEntry> abi.find(abiEntry => abiEntry.type === "fallback") || DEFAULT_FALLBACK_ABI;
   }
 
-  export function fallbackAbiForPayability(payable: boolean): FallbackAbiEntry {
+  export function fallbackAbiForPayability(payable: boolean): AbiTypes.FallbackAbiEntry {
     return {
       type: "fallback",
       stateMutability: payable ? "payable" : "nonpayable",
@@ -122,7 +77,7 @@ export namespace AbiUtils {
   }
 
   //shim for old abi versions
-  function abiMutability(abiEntry: FunctionAbiEntry | ConstructorAbiEntry | FallbackAbiEntry): Mutability {
+  function abiMutability(abiEntry: AbiTypes.FunctionAbiEntry | AbiTypes.ConstructorAbiEntry | AbiTypes.FallbackAbiEntry): Mutability {
     if(abiEntry.stateMutability !== undefined) {
       return abiEntry.stateMutability;
     }
@@ -136,16 +91,16 @@ export namespace AbiUtils {
   }
 
   //NOTE: this function returns the written out SIGNATURE, not the SELECTOR
-  export function abiSignature(abiEntry: FunctionAbiEntry | EventAbiEntry): string {
+  export function abiSignature(abiEntry: AbiTypes.FunctionAbiEntry | AbiTypes.EventAbiEntry): string {
     return abiEntry.name + abiTupleSignature(abiEntry.inputs);
   }
 
-  export function abiTupleSignature(parameters: AbiParameter[]): string {
+  export function abiTupleSignature(parameters: AbiTypes.AbiParameter[]): string {
     let components = parameters.map(abiTypeSignature);
     return "(" + components.join(",") + ")";
   }
 
-  function abiTypeSignature(parameter: AbiParameter): string {
+  function abiTypeSignature(parameter: AbiTypes.AbiParameter): string {
     let tupleMatch = parameter.type.match(/tuple(.*)/);
     if(tupleMatch === null) { //does not start with "tuple"
       return parameter.type;
@@ -157,7 +112,7 @@ export namespace AbiUtils {
     }
   }
 
-  export function abiSelector(abiEntry: FunctionAbiEntry | EventAbiEntry): string {
+  export function abiSelector(abiEntry: AbiTypes.FunctionAbiEntry | AbiTypes.EventAbiEntry): string {
     let signature = abiSignature(abiEntry);
     //NOTE: web3's soliditySha3 has a problem if the empty
     //string is passed in.  Fortunately, that should never happen here.
@@ -171,7 +126,7 @@ export namespace AbiUtils {
   }
 
   //note: undefined does not match itself :P
-  export function abisMatch(entry1: AbiEntry | undefined, entry2: AbiEntry | undefined): boolean {
+  export function abisMatch(entry1: AbiTypes.AbiEntry | undefined, entry2: AbiTypes.AbiEntry | undefined): boolean {
     //we'll consider two abi entries to match if they have the same
     //type, name (if applicable), and inputs (if applicable).
     //since there's already a signature function, we can just use that.
@@ -184,15 +139,15 @@ export namespace AbiUtils {
     switch(entry1.type) {
       case "function":
       case "event":
-        return abiSignature(entry1) === abiSignature(<FunctionAbiEntry|EventAbiEntry>entry2);
+        return abiSignature(entry1) === abiSignature(<AbiTypes.FunctionAbiEntry|AbiTypes.EventAbiEntry>entry2);
       case "constructor":
-        return abiTupleSignature(entry1.inputs) === abiTupleSignature((<ConstructorAbiEntry>entry2).inputs);
+        return abiTupleSignature(entry1.inputs) === abiTupleSignature((<AbiTypes.ConstructorAbiEntry>entry2).inputs);
       case "fallback":
         return true;
     }
   }
 
-  export function definitionMatchesAbi(abiEntry: AbiEntry, definition: AstDefinition, referenceDeclarations: AstReferences): boolean {
+  export function definitionMatchesAbi(abiEntry: AbiTypes.AbiEntry, definition: AstDefinition, referenceDeclarations: AstReferences): boolean {
     try {
       return abisMatch(abiEntry, definitionToAbi(definition, referenceDeclarations));
     }
@@ -201,7 +156,7 @@ export namespace AbiUtils {
     }
   }
 
-  export function topicsCount(abiEntry: EventAbiEntry): number {
+  export function topicsCount(abiEntry: AbiTypes.EventAbiEntry): number {
     let selectorCount = abiEntry.anonymous ? 0 : 1; //if the event is not anonymous, we must account for the selector
     return abiEntry.inputs.filter(({ indexed }) => indexed).length + selectorCount;
   }
