+++ packages/codec/lib/utils/compiler.ts
@@ -1,13 +1,8 @@
+import debugModule from "debug";
+const debug = debugModule("codec:utils:compiler");
+
 import semver from "semver";
-
-export interface CompilerVersion {
-  name?: string;
-  version?: string;
-  //NOTE: both these should really be present,
-  //but they need to be optional for compilation reasons
-}
-
-export type SolidityFamily = "pre-0.5.0" | "0.5.x";
+import { CompilerVersion, SolidityFamily } from "../types/compiler";
 
 export function solidityFamily(compiler: CompilerVersion): SolidityFamily {
   if(semver.satisfies(compiler.version, "~0.5 || >=0.5.0", {includePrerelease: true})) {
