--- packages/codec/lib/utils/errors.ts
@@ -1,11 +1,0 @@
-export class UnknownUserDefinedTypeError extends Error {
-  public typeString: string;
-  public id: string;
-  constructor(id: string, typeString: string) {
-    const message = `Cannot locate definition for ${typeString} (ID ${id})`;
-    super(message);
-    this.name = "UnknownUserDefinedTypeError";
-    this.id = id;
-    this.typeString = typeString;
-  }
-}
