+++ packages/codec/lib/utils/evm.ts
@@ -1,5 +1,5 @@
 import debugModule from "debug";
-const debug = debugModule("codec-utils:evm");
+const debug = debugModule("codec:utils:evm");
 
 import BN from "bn.js";
 import Web3 from "web3";
